backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
tiktoken
httpx
openai
anthropic
python-multipart
passlib[bcrypt]
argon2-cffi
bcrypt==3.2.2
fastapi-mail
qrcode
pillow
cryptography
prometheus-fastapi-instrumentator
psycopg2-binary
redis
slowapi
python-dotenv
